% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cw.R
\name{cw}
\alias{cw}
\title{Console wrap. Formats a string for printing to the console window}
\usage{
cw(x, final.cr = TRUE, indent = 0, exdent = 0)
}
\arguments{
\item{x}{A character object}

\item{final.cr}{Whether to add a carriage return at the end. T/F.}

\item{indent}{First line indent (number of blank spaces)}

\item{exdent}{Subsequent line(s) indent (number of blank spaces)}
}
\value{
character object with embedded \\n characters
}
\description{
Takes a potentially long character object and adds carriage returns and blank spaces for better
appearance when printing to the terminal window
}
\note{
This function can prevent long messages from extending beyond the width of the R-console.
Works with any function that prints to the console (e.g., cat, print, stop)
}
