% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo2utm.R
\name{geo2utm}
\alias{geo2utm}
\title{Look up  TM zone}
\usage{
geo2utm(x, lat = NULL)
}
\arguments{
\item{x}{Longitude in decimal degrees. Can also be a numeric vector of length 2 containing longitude and latitude values.}

\item{lat}{Latitude in decimal degrees}
}
\value{
A \code{CRS} object containing the UTM zone
}
\description{
Finds the UTM zone for a geographic coordinate
}
\details{

}
