% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.uavimg_info.R
\name{summary.uavimg_info}
\alias{summary.uavimg_info}
\title{Show a summary of a drone images metadata collection object}
\usage{
\method{summary}{uavimg_info}(object, ...)
}
\arguments{
\item{object}{A drone images metadata object}
}
\description{
Prints a summary of a drone images metadata collection object
}
\details{
Prints a summary of a drone images metadata collection object
}
\seealso{
\link{uavimg_info}
}
