% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uavimg_exp.R
\name{uavimg_exp}
\alias{uavimg_exp}
\title{Export map files from UAV images}
\usage{
uavimg_exp(x, ctr = FALSE, fp = FALSE, mcp = FALSE, shp_dir = NULL,
  create_dir = TRUE, quiet = FALSE)
}
\arguments{
\item{x}{A list of class 'uavimg_info'}

\item{ctr}{Export the image centroids as a Shapefile, T/F or a filename}

\item{fp}{Export the image footprints as a Shapefile, T/F or a filename}

\item{mcp}{Export the minimum convex polygon of the image footprints as a Shapefile, T/F or a filename}

\item{shp_dir}{The directory where the Shapefiles should be saved. If NULL, they will be saved in a 'map' sub-directory of the image folder}

\item{create_dir}{Create the output directory if it doesn't exist}

\item{quiet}{TRUE to supress printing of the pandoc command line.}
}
\value{
A vector of Shapefiles saved
}
\description{
Export map files from UAV images
}
\details{
\code{ctr}, \code{fp}, and \code{mcp} can be logical (TRUE/FALSE) or a filename (minus the shp extension).  If Logical values are pased, a Shapefile name will be constructed based on the name of parent folder. Exporting objects to Shapefile requires the \code{rgdal} package to be installed. You can specify which directory the Shapefiles will be exported to using the \code{shp_dir} argument. The default is to save them in a subdirectory called 'map' in the images directory.
}
\seealso{
\link{uavimg_info}
}
