% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uavimg_info.R
\name{uavimg_info}
\alias{uavimg_info}
\title{Extract info from UAV Images}
\usage{
uavimg_info(img_dir, exiftool = NULL, csv = NULL, alt_agl = NULL,
  fwd_overlap = TRUE, shp_save = FALSE, shp_dir = img_dir,
  shp_name = c("img_ctr", "img_fp"), quiet = FALSE)
}
\arguments{
\item{img_dir}{A directory where the image files reside}

\item{exiftool}{The path to the exiftool command line tool (omit if on the OS path)}

\item{csv}{The file name of a new csv file where the exif data will be saved (omit to make a temp one)}

\item{alt_agl}{The elevation above ground level in meters (optional for images with the relevative altitude saved)}

\item{fwd_overlap}{Whether or not to compute the amount of overlap between one image and the next, T/F}

\item{shp_save}{Save the image centroids and footprints as Shapefiles, T/F}

\item{shp_dir}{The directory where the Shapefiles should be saved}

\item{shp_name}{A character vector of the filename(s) of the image centroid and footprint Shapefiles to be saved, see Details}

\item{quiet}{Don't show messages}
}
\value{
A named list with three elements: 1) SpatialPointsDataFrame with the image centroids, and 2) a SpatialPolygonsDataFrame of the image footprints, and 3) the image directory. An HTML report of the images can be created with \link{uavimg_report}.
}
\description{
Extracts info from geotagged images taken from a drone
}
\details{
This will read the EXIF header data from a directory of image files, and plot the centroids and image footprints on the
ground. Mapping the image locations requires that the images have geostamps. In addition, mapping the image footprints
requires that the camera parameters are known, and the flight elevation about ground level is either saved in the EXIF info, or provided in the \code{alt_agl} argument. If \code{alt_agl} is passed, it will override any elevation data in the EXIF info.

This function uses a free command line tool called EXIFtool to read the EXIF data, which can be downloaded
from \url{http://www.sno.phy.queensu.ca/~phil/exiftool/}.  After you download it, rename the executable file,
'exiftool(-k).exe' to 'exiftool.exe', and save it somewhere on your system's PATH (e.g., c:\\Windows).

If \code{shp_save=TRUE}, the centroids and footprints will be exported as Shapefiles (this can also be done using the object returned by the function). If a vector of length two is passed (e.g., \code{c(T,F)}), the first and second values will determine whether to export the centroids and footprints respectively. Exporting objects to Shapefile requires the \code{rgdal} package to be installed. You can specify which directory the Shapefiles will be exported to using the \code{shp_dir} argument. The default is to save them in the same directory as the images. \code{shp_name} can be used to pass the names of the Shapefile filenames (minus the extension) for the centroids and footprints (must pass both even if you are not saving both). If \code{shp_name="dir"}, the name of directory in which the images are saved will be used as the base of the Shapefile filenames.
}
\seealso{
\link{uavimg_report}
}
