% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uavimg_move.R
\name{uavimg_move}
\alias{uavimg_move}
\title{Move UAV Images into Sub-Directories by Flight}
\usage{
uavimg_move(x, thresh_units = c("msi", "secs")[1], thresh_val = 10,
  dirnames = "", init_flt = 1, min_images = 5, preview = FALSE,
  copymove = "move", basedir = NULL, alt = NULL, quiet = FALSE)
}
\arguments{
\item{x}{A list of class 'uavimg_info'}
}
\value{
A \link{uavimg_info} object
}
\description{
Move UAV Images into Sub-Directories by Flight
}
\details{
This will move the image files into subdirectories
}
\seealso{
\link{uavimg_info}
}
