% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uavimg_exp.R
\name{uavimg_exp}
\alias{uavimg_exp}
\title{Export map files from UAV images}
\usage{
uavimg_exp(x, shp_save = TRUE, shp_dir = NULL, shp_name = "dir",
  create_dir = TRUE, quiet = FALSE)
}
\arguments{
\item{x}{A list of class 'uavimg_info'}

\item{shp_save}{Save the image centroids and footprints as Shapefiles, T/F}

\item{shp_dir}{The directory where the Shapefiles should be saved. If NULL, they will be saved in a 'map' sub-directory of the image folder}

\item{shp_name}{A character vector of the filename(s) of the image centroid and footprint Shapefiles to be saved, see Details}

\item{create_dir}{Create the output directory if it doesn't exist}

\item{quiet}{TRUE to supress printing of the pandoc command line.}
}
\value{

}
\description{
Export map files from UAV images
}
\details{
If \code{shp_save=TRUE}, the centroids and footprints will be exported as Shapefiles. If a vector of length two is passed (e.g., \code{c(T,F)}), the first and second values will determine whether to export the centroids and footprints respectively. Exporting objects to Shapefile requires the \code{rgdal} package to be installed. You can specify which directory the Shapefiles will be exported to using the \code{shp_dir} argument. The default is to save them in a subdirectory called 'map' in the images directory. \code{shp_name} can be used to pass the names of the Shapefile filenames (minus the extension) for the centroids and footprints (must pass both even if you are not saving both). If \code{shp_name="dir"}, the name of directory in which the images are saved will be used as the base of the Shapefile filenames.
}
\seealso{
\link{uavimg_info}
}
