% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uavimg_info.R
\name{uavimg_info}
\alias{uavimg_info}
\title{Extract info from UAV Images}
\usage{
uavimg_info(img_dir, exiftool = NULL, csv = NULL, alt_agl = NULL,
  fwd_overlap = TRUE, quiet = FALSE)
}
\arguments{
\item{img_dir}{A directory where the image files reside}

\item{exiftool}{The path to the exiftool command line tool (omit if on the OS path)}

\item{csv}{The file name of a new csv file where the exif data will be saved (omit to make a temp one)}

\item{alt_agl}{The elevation above ground level in meters (optional for images with the relevative altitude saved)}

\item{fwd_overlap}{Whether or not to compute the amount of overlap between one image and the next, T/F}

\item{quiet}{Don't show messages}
}
\value{
A named list with three elements: 1) SpatialPointsDataFrame with the image centroids, and 2) a SpatialPolygonsDataFrame of the image footprints, and 3) the image directory. An HTML report of the images can be created with \link{uavimg_report}.
}
\description{
Extracts info from geotagged images taken from a drone
}
\details{
This will read the EXIF header data from a directory of image files, and plot the centroids and image footprints on the
ground. Mapping the image locations requires that the images have geostamps. In addition, mapping the image footprints
requires that the camera parameters are known, and the flight elevation about ground level is either saved in the EXIF info, or provided in the \code{alt_agl} argument. If \code{alt_agl} is passed, it will override any elevation data in the EXIF info.

This function uses a free command line tool called EXIFtool to read the EXIF data, which can be downloaded
from \url{http://www.sno.phy.queensu.ca/~phil/exiftool/}.  After you download it, rename the executable file,
'exiftool(-k).exe' to 'exiftool.exe', and save it somewhere on your system's PATH (e.g., c:\\Windows).
}
\seealso{
\link{uavimg_report}
}
