% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uavimg_report.R
\name{uavimg_report}
\alias{uavimg_report}
\title{Create report of UAV Images}
\usage{
uavimg_report(x, output_file = NULL, output_dir = NULL, create_dir = TRUE,
  report_rmd = NULL, open_report = TRUE, self_contained = TRUE,
  quiet = FALSE)
}
\arguments{
\item{x}{A list of class 'uavimg_info'}

\item{output_file}{If NULL then a default based on the name of the input file is chosen.}

\item{output_dir}{If NULL, then will be placed in a 'map' sub-directory of the images}

\item{create_dir}{Create the output directory if it doesn't exist}

\item{report_rmd}{RMarkdown filename for the report}

\item{open_report}{Open the HTML file in a browser}

\item{self_contained}{Make the output HTML file self-contained}

\item{quiet}{TRUE to supress printing of the pandoc command line.}
}
\value{
The filename of the HTML report generated
}
\description{
Creates a report of UAV images
}
\details{

}
\seealso{
\link{uavimg_info}
}
